+++ infrastructure/database/models/referrals.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Integer, String
+from sqlalchemy import String, BigInteger, Numeric
 from sqlalchemy.orm import Mapped, mapped_column
 
 from .base import BaseModel
@@ -7,7 +7,7 @@ from .base import BaseModel
 class Referrals(BaseModel):
     __tablename__ = "referrals"
 
-    user_id: Mapped[int] = mapped_column(Integer)
-    referral_user_id: Mapped[int] = mapped_column(Integer)
+    user_id: Mapped[int] = mapped_column(BigInteger)
+    referral_user_id: Mapped[int] = mapped_column(BigInteger)
     referral_username: Mapped[str] = mapped_column(String)
-    referral_reg_data: Mapped[int] = mapped_column(Integer)
+    referral_reg_data: Mapped[int] = mapped_column(Numeric)
