+++ infrastructure/database/models/user_services.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Integer, ForeignKey
+from sqlalchemy import Integer, ForeignKey, BigInteger
 from sqlalchemy.orm import Mapped, mapped_column
 
 from infrastructure.database.models import BaseModel
@@ -7,6 +7,6 @@ from infrastructure.database.models import BaseModel
 class UserServices(BaseModel):
     __tablename__ = "user_services"
 
-    user_id: Mapped[int] = mapped_column(Integer)
+    user_id: Mapped[int] = mapped_column(BigInteger)
     service: Mapped[int] = mapped_column(Integer)
     token_id: Mapped[int] = mapped_column(Integer, ForeignKey("user_tokens.id"))
